+++ layers/src/main/java/dao/CakeToppingDao.java
@@ -16,25 +16,24 @@
  *
  * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
  * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * FITNESS FOR A PARTICULAR PURPOSE AND NON INFRINGEMENT. IN NO EVENT SHALL THE
  * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
  * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package com.iluwatar.layers.exception;
+package dao;
+
+
+
+import entity.CakeTopping;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
 
 /**
- * Custom exception used in cake baking.
+ * CRUD repository cake toppings.
  */
-public class CakeBakingException extends Exception {
+@Repository
+public interface CakeToppingDao extends JpaRepository<CakeTopping, Long> {
 
-  private static final long serialVersionUID = 1L;
-
-  public CakeBakingException() {
-  }
-
-  public CakeBakingException(String message) {
-    super(message);
-  }
 }
