+++ layers/src/main/java/entity/CakeLayer.java
@@ -22,13 +22,75 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package com.iluwatar.layers.view;
+package entity;
+
+
+
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.Id;
+import jakarta.persistence.ManyToOne;
+import jakarta.persistence.CascadeType;
 
 /**
- * View interface.
+ * CakeLayer entity.
  */
-public interface View {
+@Entity
+public class CakeLayer {
 
-  void render();
+    @Id
+    @GeneratedValue
+    private Long id;
 
+    private String name;
+
+    private int calories;
+
+    @ManyToOne(cascade = CascadeType.ALL)
+    private Cake cake;
+
+    public CakeLayer() {
+    }
+
+    public CakeLayer(String name, int calories) {
+        this.setName(name);
+        this.setCalories(calories);
+    }
+
+    public Long getId() {
+        return id;
+    }
+
+    public void setId(Long id) {
+        this.id = id;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
+    public int getCalories() {
+        return calories;
+    }
+
+    public void setCalories(int calories) {
+        this.calories = calories;
+    }
+
+    @Override
+    public String toString() {
+        return String.format("id=%s name=%s calories=%d", id, name, calories);
+    }
+
+    public Cake getCake() {
+        return cake;
+    }
+
+    public void setCake(Cake cake) {
+        this.cake = cake;
+    }
 }
