+++ layers/src/main/java/entity/CakeTopping.java
@@ -22,13 +22,14 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package com.iluwatar.layers.entity;
+package entity;
 
-import javax.persistence.CascadeType;
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.Id;
-import javax.persistence.OneToOne;
+
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.Id;
+import jakarta.persistence.OneToOne;
+import jakarta.persistence.CascadeType;
 
 /**
  * CakeTopping entity.
@@ -36,59 +37,59 @@ import javax.persistence.OneToOne;
 @Entity
 public class CakeTopping {
 
-  @Id
-  @GeneratedValue
-  private Long id;
+    @Id
+    @GeneratedValue
+    private Long id;
 
-  private String name;
+    private String name;
 
-  private int calories;
+    private int calories;
 
-  @OneToOne(cascade = CascadeType.ALL)
-  private Cake cake;
+    @OneToOne(cascade = CascadeType.ALL)
+    private Cake cake;
 
-  public CakeTopping() {
-  }
+    public CakeTopping() {
+    }
 
-  public CakeTopping(String name, int calories) {
-    this.setName(name);
-    this.setCalories(calories);
-  }
+    public CakeTopping(String name, int calories) {
+        this.setName(name);
+        this.setCalories(calories);
+    }
 
-  public Long getId() {
-    return id;
-  }
+    public Long getId() {
+        return id;
+    }
 
-  public void setId(Long id) {
-    this.id = id;
-  }
+    public void setId(Long id) {
+        this.id = id;
+    }
 
-  public String getName() {
-    return name;
-  }
+    public String getName() {
+        return name;
+    }
 
-  public void setName(String name) {
-    this.name = name;
-  }
+    public void setName(String name) {
+        this.name = name;
+    }
 
-  public int getCalories() {
-    return calories;
-  }
+    public int getCalories() {
+        return calories;
+    }
 
-  public void setCalories(int calories) {
-    this.calories = calories;
-  }
+    public void setCalories(int calories) {
+        this.calories = calories;
+    }
 
-  @Override
-  public String toString() {
-    return String.format("id=%s name=%s calories=%d", id, name, calories);
-  }
+    @Override
+    public String toString() {
+        return String.format("id=%s name=%s calories=%d", id, name, calories);
+    }
 
-  public Cake getCake() {
-    return cake;
-  }
+    public Cake getCake() {
+        return cake;
+    }
 
-  public void setCake(Cake cake) {
-    this.cake = cake;
-  }
+    public void setCake(Cake cake) {
+        this.cake = cake;
+    }
 }
