+++ layers/src/main/resources/META-INF/logback.xml
@@ -16,7 +16,7 @@
 
     THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
     IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-    FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+    FITNESS FOR A PARTICULAR PURPOSE AND NON INFRINGEMENT. IN NO EVENT SHALL THE
     AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
     LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
     OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
@@ -24,30 +24,30 @@
 
 -->
 <configuration>
-  <appender name="FILE" class="ch.qos.logback.core.rolling.RollingFileAppender">
-    <file>layers.log</file>
-    <rollingPolicy class="ch.qos.logback.core.rolling.TimeBasedRollingPolicy">
-      <fileNamePattern>layers-%d.log</fileNamePattern>
-      <maxHistory>5</maxHistory>
-    </rollingPolicy>
-    <encoder>
-      <pattern>%-5p [%d{ISO8601,UTC}] %c: %m%n</pattern>
-    </encoder>
-  </appender>
-  <appender name="STDOUT" class="ch.qos.logback.core.ConsoleAppender">
-    <encoder>
-      <pattern>%-5p [%d{ISO8601,UTC}] %c: %m%n</pattern>
-    </encoder>
-  </appender>
-  <logger name="com.iluwatar" additivity="false">
-    <level value="DEBUG" />
-    <appender-ref ref="FILE" />
-  </logger>
-  <logger name="org.hibernate" additivity="false">
-    <level value="ERROR" />
-    <appender-ref ref="STDOUT" />
-  </logger>
-  <root level="WARN">
-    <appender-ref ref="STDOUT" />
-  </root>
+    <appender name="FILE" class="ch.qos.logback.core.rolling.RollingFileAppender">
+        <file>layers.log</file>
+        <rollingPolicy class="ch.qos.logback.core.rolling.TimeBasedRollingPolicy">
+            <fileNamePattern>layers-%d.log</fileNamePattern>
+            <maxHistory>5</maxHistory>
+        </rollingPolicy>
+        <encoder>
+            <pattern>%-5p [%d{ISO8601,UTC}] %c: %m%n</pattern>
+        </encoder>
+    </appender>
+    <appender name="STDOUT" class="ch.qos.logback.core.ConsoleAppender">
+        <encoder>
+            <pattern>%-5p [%d{ISO8601,UTC}] %c: %m%n</pattern>
+        </encoder>
+    </appender>
+    <logger name="com.iluwatar" additivity="false">
+        <level value="DEBUG" />
+        <appender-ref ref="FILE" />
+    </logger>
+    <logger name="org.hibernate" additivity="false">
+        <level value="ERROR" />
+        <appender-ref ref="STDOUT" />
+    </logger>
+    <root level="WARN">
+        <appender-ref ref="STDOUT" />
+    </root>
 </configuration>
