+++ layers/src/main/resources/application.properties
@@ -1,0 +1,12 @@
+
+spring.main.web-application-type=none
+
+#datasource settings
+spring.datasource.url=jdbc:h2:~/databases/cake
+spring.datasource.driverClassName=org.h2.Driver
+spring.datasource.username=sa
+spring.datasource.password=sa
+
+#data settings
+spring.jpa.database-platform=org.hibernate.dialect.H2Dialect
+spring.jpa.hibernate.ddl-auto=create-drop
