+++ layers/src/test/java/com/iluwatar/layers/app/LayersAppTests.java
@@ -1,0 +1,13 @@
+package com.iluwatar.layers.app;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest(classes = LayersApp.class)
+class LayersAppTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
